pyproject.toml
[project]
name = "leddisplay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=16.1.0",
    "cyndilib>=0.1.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.1.0",
]
